Don't send events in DEVICE_INIT, only after DEVICE_ON.
See http://bugs.freedesktop.org/show_bug.cgi?id=2243#c15 for a description of
the problem fixed in this patch.

Now only the driver leds state is synched with virtual console state, and if
required, events are post in DEVICE_ON.
This also defines an internal INITFLAG state to know when DEVICE_ON is called
after DEVICE_INIT, otherwise, it will only send events for leds that changed
state while the device was in "DEVICE_OFF state".

Signed-off-by: Peter Hutterer <peter@cs.unisa.edu.au>